Add support for return statements in functions
- Extend AST and parser to handle 'return' statements with optional value
- Implement runtime handling of return in interpreter using ReturnObj
- Update function call logic in LoxFunction to properly catch and process returns
- Add example files demonstrating return, clock, and fib usage
- Ensure clock() returns a float value for consistency in arithmetic

+++ lox/clock.go
@@ -11,5 +11,5 @@ func (c ClockNativeFunction) Arity() int {
 }
 
 func (c ClockNativeFunction) Call(i *Interpreter, arguments []any) (any, LoxError) {
-	return time.Now().UnixMilli(), nil
+	return float64(time.Now().UnixMilli()), nil
 }
